Design: 헤더 UI 변경
