--- src/components/layout/header/header.module.scss
@@ -1,132 +1,0 @@
-.container {
-  width: 100%;
-  height: $--header-height;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 5rem;
-  margin-top: 2rem;
-  font-weight: 600;
-
-  .logo {
-    text-transform: uppercase;
-    cursor: pointer;
-    img {
-      width: 180px;
-      height: auto;
-      margin-bottom: 32px;
-    }
-  }
-  .nav {
-    display: flex;
-  }
-  .nav > div {
-    text-decoration: none;
-    color: $black;
-    cursor: pointer;
-    padding: 1rem 2.5rem;
-    border-bottom: 2px solid transparent;
-
-    a {
-      color: #b5b5b5;
-      text-decoration: none;
-    }
-
-    &:hover {
-      box-sizing: border-box;
-      border-bottom: 2px solid $red-500;
-      color: $black;
-    }
-
-    .nav-link {
-      color: $black;
-    }
-  }
-
-  .signin {
-    text-transform: capitalize;
-    font-size: 16px;
-    color: #ed224b;
-    background-color: transparent;
-    outline: none;
-    display: flex;
-    flex-direction: row;
-    column-gap: 1rem;
-
-    a {
-      text-decoration: none;
-      color: $red-500;
-      font-weight: bold;
-    }
-
-    button {
-      color: $red-500;
-      font-weight: bold;
-      background-color: transparent;
-      font-size: medium;
-    }
-  }
-}
-
-@include mobile {
-  .menu-btn {
-    flex: 1.5;
-    text-align: center;
-    cursor: pointer;
-    font-size: 1.5rem;
-  }
-
-  .signin {
-    flex: 2;
-    text-align: center;
-    padding: 0 0.5rem 0 0;
-  }
-
-  .logo {
-    flex: 8;
-    text-align: center;
-  }
-
-  .nav {
-    visibility: hidden;
-  }
-
-  .nav > div {
-    display: none;
-  }
-}
-
-@include tablet {
-  .menu-btn {
-    flex: 1;
-    text-align: center;
-    cursor: pointer;
-    font-size: 1.5rem;
-  }
-
-  .signin {
-    flex: 1.5;
-    text-align: center;
-  }
-
-  .logo {
-    flex: 8;
-    text-align: center;
-  }
-
-  .nav {
-    visibility: hidden;
-  }
-
-  .nav > div {
-    display: none;
-  }
-}
-@include desktop {
-  .container {
-    column-gap: 3rem;
-  }
-  .menu-btn {
-    display: none;
-  }
-}
