+++ src/components/layout/header/index.module.scss
@@ -55,6 +55,7 @@
         opacity: 0.5;
         font-size: 14px;
         padding: 0.75rem;
+        cursor: pointer;
 
         &:hover {
           opacity: 1;
